requirements.txt
bitso-py==3.0.1
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
idna==2.10
mock==4.0.3
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
websocket-client==0.40.0
